requirements.txt
langchain>=0.1.16
langchain-core>=0.1.33
langchain-community>=0.0.29
langchain-openai>=0.0.8
langgraph==0.2.53
openai==1.55.3
pytesseract==0.3.10
pillow==10.2.0
opencv-python==4.9.0.80
openai-whisper==20231117
ffmpeg-python==0.2.0
python-dotenv==1.0.1
pydantic==2.6.1
loguru==0.7.2
fastapi==0.110.0
python-multipart==0.0.9
uvicorn==0.27.1
easyocr==1.7.1